SpringFundAPI/requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-mongodb-backend==5.0.0b1
pymongo==4.13.2
djangorestframework
django-cors-headers
